Added a complete, unedited transcript
This file is a complete session using chatgpt_cli. The program keeps a complete transcript for two purposes. One is to allow loading of the previous conversation for purposes of continuing a coherent discussion. The other is to load an older conversation for the same purposes. The server does not maintain context, so without context information, each message sent to OpenAI's chat (GPT-4, GPT-3.5turbo, etc) is as if you approached it for the very first time with a prompt.

I decided to keep this one in raw text format. Some strange characters may appear if UTF-8 or Latin-1 encoding was used.